add factory & seeder

+++ database/migrations/2024_09_21_045747_create_patients_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('patients', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->unsignedBigInteger('number')->unique();
+            $table->integer('number')->unique();
             $table->timestamps();
         });
     }
